+++ src/Bio/HTS.hs
@@ -1,0 +1,7 @@
+module Bio.HTS
+    ( module Bio.HTS.Types
+    , module Bio.HTS.BAM
+    ) where
+
+import Bio.HTS.Types
+import Bio.HTS.BAM