+++ ChangeLog.md
@@ -1,5 +1,12 @@
 # Revision history for HsHTSLib
 
-## 1.9.0  -- YYYY-mm-dd
+## 1.9.2  -- 2019-03-18
 
-* htslib-1.9
+* Name changes: seqName -> queryName, getChrId -> refId, getChr -> refName,
+mateChrId -> mateRefId, mateChr -> mateRefName.
+
+* Use 1-based coordinate system in SAM format.
+
+## 1.9.1  -- 2019-03-18
+
+* Upgrade the C library to v1.9.